Add complete JWT authentication

+++ backend/src/main/java/mj/project/stickynotesscrubbackend/app_user/entity/AppUser.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import mj.project.stickynotesscrubbackend.notes.entity.Note;
+import mj.project.stickynotesscrubbackend.note.entity.Note;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
 
@@ -22,7 +22,7 @@ public class AppUser implements UserDetails {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int id;
+    private long id;
 
     @Column(unique = true)
     private String username;
